move arraysize function to sequences
